do not regenerate the runfile if it exists

+++ src/bret/evaluation.py
@@ -1,5 +1,6 @@
 import json
 import logging
+import os
 import time
 
 import numpy as np
@@ -11,22 +12,27 @@ logger = logging.getLogger(__name__)
 
 
 class Evaluator:
-    def __init__(self, model, index, device, metrics={"ndcg", "map", "recip_rank"}):
+    def __init__(self, model, device, index=None, metrics={"ndcg", "map", "recip_rank"}):
         self.model = model
-        self.index = index
         self.device = device
+        self.index = index
         self.metrics = metrics
 
     def evaluate_retriever(self, qry_data_loader, qrels, k=10, num_samples=None, run_file=None):
-        logger.info("Generating run...")
-        t_start = time.time()
-        run = self._generate_run(qry_data_loader, k=k, num_samples=num_samples)
-        t_end = time.time()
-        logger.info("Run generated in %.2f minutes.", (t_end - t_start) / 60)
-        if run_file is not None:
-            with open(run_file, "w", encoding="utf-8") as f:
-                json.dump(run, f)
-            logger.info("Run saved in: %s", run_file)
+        if os.path.exists(run_file) and os.path.isfile(run_file):
+            logger.info("Loading run from: %s", run_file)
+            with open(run_file, "r", encoding="utf-8") as f:
+                run = json.loads(f.read())
+        else:
+            logger.info("Generating run...")
+            t_start = time.time()
+            run = self._generate_run(qry_data_loader, k=k, num_samples=num_samples)
+            t_end = time.time()
+            logger.info("Run generated in %.2f minutes.", (t_end - t_start) / 60)
+            if run_file is not None:
+                with open(run_file, "w", encoding="utf-8") as f:
+                    json.dump(run, f)
+                logger.info("Run saved in: %s", run_file)
         logger.info("Calculating metrics...")
         results = self._calculate_metrics(run, qrels, k=k)
         return results
